requirements.txt
requests
psutil
appdirs
python-telegram-bot>=13.0,<14.0
google-api-python-client>=1.12.8,<2.0.0
google-auth-httplib2>=0.0.3,<0.1.0
google-auth-oauthlib>=0.4.1,<0.10.0
aria2p>=0.9.0,<0.15.0
python-dotenv>=0.10
tenacity>=6.0.0
python-magic
beautifulsoup4>=4.8.2,<4.8.10
Pyrogram>=0.16.0,<0.16.10
TgCrypto>=1.1.1,<1.1.10
youtube_dl
lxml
telegraph
